Make "pick" helper types less restrictive
In our codebase we frequently use the pick helper with objects that aren't Events (although that is one of the more common use cases). For example we have generic components that fire an action with an object, and we only want one property from the object to be passed up the action chain to more specific components. This attempts to relax the typing on the pick helper so it will accept either an object (or rather `Record<string, unknown>`) OR an Event.